# Import/Perf: pick low-hanging performance fruit for big meshes