--- a/addons/io_scene_gltf2/blender/imp/gltf2_blender_primitive.py
+++ b/addons/io_scene_gltf2/blender/imp/gltf2_blender_primitive.py
@@ -140,15 +140,15 @@ class BlenderPrimitive():
                 )
 
             for bidx, pidx in vert_idxs:
+                color = colors[pidx]
+                col = (
+                    color_linear_to_srgb(color[0]),
+                    color_linear_to_srgb(color[1]),
+                    color_linear_to_srgb(color[2]),
+                    color[3] if is_rgba else 1.0,
+                )[:blender_num_components]
                 for loop in bme_verts[bidx].link_loops:
-                    color = colors[pidx]
-                    col = (
-                        color_linear_to_srgb(color[0]),
-                        color_linear_to_srgb(color[1]),
-                        color_linear_to_srgb(color[2]),
-                        color[3] if is_rgba else 1.0,
-                    )
-                    loop[layer] = col[:blender_num_components]
+                    loop[layer] = col
 
             set_num += 1