So I tried importing the model from [T72167: glTF file loads extremely slowly (5+ min)](https://developer.blender.org/T72167). Seems a lot of time is spent in mesh creation, which is pretty slow for big models, but there was some low-hanging fruit that was really easy to pick. This knocks off about 40% of the import time for me.

| Shading option | before | after |
|:---|:---|:---|
| Use Normal Data | 4 min 6 sec | 2 min 30 sec |
| Smooth Shading | 3 min 12 sec | 1 min 38 sec |

The difference is pretty much undetectable in small models.