@scurest 2020-01-23
Also helps MetalRoughSpheres

| Shading option | before | after |
|:---|:---|:---|
| Use Normal Data | 5.1s | 4.0s |
| Smooth Shading | 3.5s | 2.5s |