@julienduroure 2020-01-23
Funny fact:
You commented using foreach_set (in #76) one year ago, exaclty (Januray 23rd 2019 / January 23rd 2020)

Thanks @scurest !